requirements.txt
rembg==2.0.30
Pillow==9.3.0
pocketbase==0.8.0